Move plugin code to plugin.go
